+++ benchmarks/Classifiers/KNearestNeighborsBench.php
@@ -6,11 +6,14 @@ use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Classifiers\KNearestNeighbors;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 
+/**
+ * @Groups({"Classifiers"})
+ */
 class KNearestNeighborsBench
 {
     protected const TRAINING_SIZE = 2500;
 
-    protected const TESTING_SIZE = 5000;
+    protected const TESTING_SIZE = 10000;
 
     /**
      * @var \Rubix\ML\Datasets\Labeled;
@@ -30,10 +33,10 @@ class KNearestNeighborsBench
     public function setUpTrainPredict() : void
     {
         $generator = new Agglomerate([
-            'red' => new Blob([255, 32, 0], 30.),
-            'green' => new Blob([0, 128, 0], 10.),
-            'blue' => new Blob([0, 32, 255], 20.),
-        ], [2, 3, 4]);
+            'Iris-setosa' => new Blob([5.0, 3.42, 1.46, 0.24], [0.35, 0.38, 0.17, 0.1]),
+            'Iris-versicolor' => new Blob([5.94, 2.77, 4.26, 1.33], [0.51, 0.31, 0.47, 0.2]),
+            'Iris-virginica' => new Blob([6.59, 2.97, 5.55, 2.03], [0.63, 0.32, 0.55, 0.27]),
+        ]);
 
         $this->training = $generator->generate(self::TRAINING_SIZE);
 
@@ -43,7 +46,9 @@ class KNearestNeighborsBench
     }
 
     /**
+     * @Iterations(1)
      * @BeforeMethods({"setUpTrainPredict"})
+     * @OutputTimeUnit("seconds", precision=3)
      */
     public function bench_train_predict() : void
     {