+++ benchmarks/Classifiers/MultilayerPerceptronBench.php
@@ -2,14 +2,17 @@
 
 namespace Rubix\ML\Benchmarks\Classifiers;
 
-use Rubix\ML\Classifiers\KDNeighbors;
+use Rubix\ML\NeuralNet\Layers\Dense;
 use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\NeuralNet\Layers\Activation;
+use Rubix\ML\Classifiers\MultilayerPerceptron;
+use Rubix\ML\NeuralNet\ActivationFunctions\ReLU;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 
 /**
  * @Groups({"Classifiers"})
  */
-class KDNeighborsBench
+class MultilayerPerceptronBench
 {
     protected const TRAINING_SIZE = 2500;
 
@@ -26,7 +29,7 @@ class KDNeighborsBench
     public $testing;
 
     /**
-     * @var \Rubix\ML\Classifiers\KNearestNeighbors
+     * @var \Rubix\ML\Classifiers\MultilayerPerceptron
      */
     protected $estimator;
 
@@ -34,7 +37,6 @@ class KDNeighborsBench
     {
         $generator = new Agglomerate([
             'Iris-setosa' => new Blob([5.0, 3.42, 1.46, 0.24], [0.35, 0.38, 0.17, 0.1]),
-            'Iris-versicolor' => new Blob([5.94, 2.77, 4.26, 1.33], [0.51, 0.31, 0.47, 0.2]),
             'Iris-virginica' => new Blob([6.59, 2.97, 5.55, 2.03], [0.63, 0.32, 0.55, 0.27]),
         ]);
 
@@ -42,7 +44,10 @@ class KDNeighborsBench
 
         $this->testing = $generator->generate(self::TESTING_SIZE);
 
-        $this->estimator = new KDNeighbors(5, true);
+        $this->estimator = new MultilayerPerceptron([
+            new Dense(100),
+            new Activation(new ReLU()),
+        ]);
     }
 
     /**