+++ benchmarks/Classifiers/RadiusNeighborsBench.php
@@ -2,14 +2,14 @@
 
 namespace Rubix\ML\Benchmarks\Classifiers;
 
-use Rubix\ML\Classifiers\KDNeighbors;
 use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\Classifiers\RadiusNeighbors;
 use Rubix\ML\Datasets\Generators\Agglomerate;
 
 /**
  * @Groups({"Classifiers"})
  */
-class KDNeighborsBench
+class RadiusNeighborsBench
 {
     protected const TRAINING_SIZE = 2500;
 
@@ -26,7 +26,7 @@ class KDNeighborsBench
     public $testing;
 
     /**
-     * @var \Rubix\ML\Classifiers\KNearestNeighbors
+     * @var \Rubix\ML\Classifiers\RadiusNeighbors
      */
     protected $estimator;
 
@@ -42,7 +42,7 @@ class KDNeighborsBench
 
         $this->testing = $generator->generate(self::TESTING_SIZE);
 
-        $this->estimator = new KDNeighbors(5, true);
+        $this->estimator = new RadiusNeighbors(0.1);
     }
 
     /**